itest: move testFundingPersistence to lnd_funding_test.go
